+++ custom_components/homematicip_local/button.py
@@ -12,6 +12,7 @@ from homeassistant.components.button import ButtonEntity
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.typing import UndefinedType
 
 from . import HomematicConfigEntry
 from .control_unit import ControlUnit, signal_new_data_point
@@ -96,9 +97,16 @@ class HaHomematicProgramButton(HaHomematicGenericHubEntity, ButtonEntity):
         )
         self._data_point: ProgramDpButton = data_point
         self._attr_extra_state_attributes = {
-            ATTR_NAME: self._data_point.ccu_program_name,
+            ATTR_NAME: self._data_point.name,
             ATTR_DESCRIPTION: self._data_point.description,
         }
+
+    @property
+    def name(self) -> str | UndefinedType | None:
+        """Return the name of the sysvar data_point."""
+        if (name := super().name) is not None and isinstance(name, str):
+            return f"p {name}".title()
+        return name
 
     async def async_press(self) -> None:
         """Execute a button press."""