+++ tests/const.py
@@ -19,7 +19,7 @@ USERNAME = "user"
 SYSVAR_DATA: list[SystemVariableData] = [
     SystemVariableData(
         vid="1",
-        name="sv_alarm",
+        name="alarm",
         description="",
         data_type=SysvarType.ALARM,
         unit=None,
@@ -31,7 +31,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="2",
-        name="sv_alarm_ext",
+        name="alarm_ext",
         description="",
         data_type=SysvarType.ALARM,
         unit=None,
@@ -43,7 +43,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="3",
-        name="sv_logic",
+        name="logic",
         description="",
         data_type=SysvarType.LOGIC,
         unit=None,
@@ -55,7 +55,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="4",
-        name="sv_logic_ext",
+        name="logic_ext",
         description="",
         data_type=SysvarType.LOGIC,
         unit=None,
@@ -67,7 +67,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="5",
-        name="sv_list",
+        name="list",
         description="",
         data_type=SysvarType.LIST,
         unit=None,
@@ -79,7 +79,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="6",
-        name="sv_list_ext",
+        name="list_ext",
         description="",
         data_type=SysvarType.LIST,
         unit=None,
@@ -91,7 +91,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="7",
-        name="sv_string",
+        name="string",
         description="",
         data_type=SysvarType.STRING,
         unit=None,
@@ -103,7 +103,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="8",
-        name="sv_string_ext",
+        name="string_ext",
         description="",
         data_type=SysvarType.STRING,
         unit=None,
@@ -115,7 +115,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="9",
-        name="sv_float",
+        name="float",
         description="",
         data_type=SysvarType.FLOAT,
         unit=None,
@@ -127,7 +127,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="10",
-        name="sv_float_ext",
+        name="float_ext",
         description="",
         data_type=SysvarType.FLOAT,
         unit="°C",
@@ -139,7 +139,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="11",
-        name="sv_integer",
+        name="integer",
         description="",
         data_type=SysvarType.INTEGER,
         unit=None,
@@ -151,7 +151,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="12",
-        name="sv_integer_ext",
+        name="integer_ext",
         description="",
         data_type=SysvarType.INTEGER,
         unit=None,